+++ src/constants.ts
@@ -7,9 +7,9 @@ export const SYS_MODE_WEBBROWSER = 'wb'
 export const SYS_MODE_NODEJS = 'njs'
 export const SYS_MODE_UNKNOWN = 'unknown'
 
-export const MODE_CONSOLE = 'console'
-export const MODE_HTML = 'html'
-export const DEFAULT_MODE = MODE_CONSOLE
+export const OUTPUT_CONSOLE = 'console'
+export const OUTPUT_HTML = 'html'
+export const DEFAULT_OUTPUT = OUTPUT_CONSOLE
 
 export const ANONYMOUS_TEST_NAME = '(anonymous)'
 export const ANONYMOUS_TEST_IDX = 0