+++ src/index.ts
@@ -1,5 +1,5 @@
 import { UserConfig, Config } from './types'
-import { DEFAULT_MODE, ANONYMOUS_TEST_NAME, ANONYMOUS_TEST_IDX } from './constants'
+import { DEFAULT_OUTPUT, ANONYMOUS_TEST_NAME, ANONYMOUS_TEST_IDX } from './constants'
 import JPerf from './JPerf'
 import { validConfig } from './validator'
 
@@ -11,7 +11,7 @@ const fillConfigDefaults = (config: UserConfig): Config => {
     hardwareDetails: conf.hardwareDetails !== undefined ? config.hardwareDetails : false,
     anonymousTestName: conf.anonymousTestName !== undefined ? config.anonymousTestName : ANONYMOUS_TEST_NAME,
     anonymousTestIndex: conf.anonymousTestIndex !== undefined ? config.anonymousTestIndex : ANONYMOUS_TEST_IDX,
-    mode: conf.mode || DEFAULT_MODE,
+    output: conf.output || DEFAULT_OUTPUT,
     selector: conf.selector || undefined
   }
 }